Delete duplicate test function in examples_complete_test.go
